% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileNames.R
\docType{data}
\name{FileNames}
\alias{FileNames}
\title{File names and corresponding ExperimentHub (EH) IDs for all 84 .RDS files
included in preciseTADhub}
\format{
An object of class \code{data.frame}
}
\usage{
data(FileNames)
}
\value{
A data.frame
}
\description{
A data frame with 2 columns and 84 rows
\describe{
\item{FileNames}{The file names for the pre-trained random forest models}
\item{EHID}{The ExperimentHub ID}
}
}
\keyword{datasets}
